+++ core/src/com/afyber/game/api/overworld/Collision.java
@@ -18,8 +18,8 @@ public class Collision extends WorldObject {
         // none
     }
 
-    public boolean isColliding(int x, int y, int width, int height) {
-        // FIXME: you can get 0.5 pixels close to right and up walls but only 1 pixel close to left and down walls
+    public boolean isColliding(float x, float y, float width, float height) {
+        // well at least it's consistent now
         if (worldPos[0] + collisionDimensions[0] >= x && x + width >= worldPos[0]) {
             if (worldPos[1] + collisionDimensions[1] >= y && y + height >= worldPos[1]) {
                 return true;
